refactor(compliance): split evidence_drop's CLI arg-parsing into a sibling module
shared/scripts/lib/evidence_drop.py crossed the 300-line guideline (314) after
the Stage-3 doubt-review fix added _missing_named_sources -- a new crossing,
blocked by the bloat anti-ratchet Stop hook (never just raise the baseline cap).

Extracted the pure CLI arg-parsing/validation leg (parse_junit_args,
missing_named_sources -- no filesystem writes, no provenance/staging logic)
into shared/scripts/lib/_evidence_drop_cli.py, mirroring this repo's existing
private-sibling convention (_backfill_fold.py / _backfill_title_sim.py beside
backfill_signals.py) including the dual-form import (flat when evidence_drop.py
is executed directly as a script, relative when imported as lib.evidence_drop).

Pure internal split, zero behavior change: the public API (stage_reports,
clear_evidence_reports, evidence_is_fresh, etc.) is unchanged and still
importable from evidence_drop exactly as before, so no consumer (compliance
collectors, verifiers, run_full_suite_evidence.py) needed an import change.
main() is untouched -- _parse_junit_args/_missing_named_sources are re-exported
under their original private names.

evidence_drop.py: 271 lines. _evidence_drop_cli.py: 77 lines. Both comfortably
under the 300-line guideline. anti_ratchet_check.py --staged: exit 0.

Run-ID: iterate-2026-08-25-r1a-evidence-staging-multiroot
